Rename PivotAssist -> PivotInput
