+++ editor/src/messages/layout/layout_message_handler.rs
@@ -196,10 +196,10 @@ impl<F: Fn(&MessageDiscriminant) -> Vec<KeysGroup>> MessageHandler<LayoutMessage
 						let callback_message = (parameter_expose_button.on_update.callback)(parameter_expose_button);
 						responses.add(callback_message);
 					}
-					Widget::PivotAssist(pivot_assist) => {
-						let update_value = value.as_str().expect("RadioInput update was not of type: u64");
-						pivot_assist.position = update_value.into();
-						let callback_message = (pivot_assist.on_update.callback)(pivot_assist);
+					Widget::PivotInput(pivot_input) => {
+						let update_value = value.as_str().expect("PivotInput update was not of type: u64");
+						pivot_input.position = update_value.into();
+						let callback_message = (pivot_input.on_update.callback)(pivot_input);
 						responses.add(callback_message);
 					}
 					Widget::PopoverButton(_) => {}
