+++ editor/src/messages/layout/utility_types/mod.rs
@@ -3,7 +3,6 @@ pub mod widgets;
 
 pub mod widget_prelude {
 	pub use super::layout_widget::*;
-	pub use super::widgets::assist_widgets::*;
 	pub use super::widgets::button_widgets::*;
 	pub use super::widgets::input_widgets::*;
 	pub use super::widgets::label_widgets::*;
