+++ editor/src/messages/layout/utility_types/widgets/mod.rs
@@ -1,4 +1,3 @@
-pub mod assist_widgets;
 pub mod button_widgets;
 pub mod input_widgets;
 pub mod label_widgets;
